services/device-shadow/requirements.txt
paho-mqtt==1.6.1
redis==5.0.1
python-dotenv==1.0.0
